MainMenu was added
