+++ core/src/com/fruits/game/MainMenu.java
@@ -3,23 +3,37 @@ package com.fruits.game;
 import com.badlogic.gdx.Gdx;
 import com.badlogic.gdx.Screen;
 import com.badlogic.gdx.graphics.OrthographicCamera;
+import com.badlogic.gdx.utils.ScreenUtils;
 
 public class MainMenu implements Screen {
-    final Fruit game;
+    final Fruits game;
 
     OrthographicCamera camera;
 
-    public MainMenu(final Fruit game) {
+    public MainMenu(final Fruits game) {
         this.game = game;
 
         camera = new OrthographicCamera();
-        camera.setToOrtho(false, 800, 480);
+        camera.setToOrtho(false, Gdx.graphics.getWidth(), Gdx.graphics.getHeight());
+    }
+
+    public void render(float delta) {
+        ScreenUtils.clear(0, 0, 0.2f, 1);
+
+        camera.update();
+        game.batch.setProjectionMatrix(camera.combined);
+
+        game.batch.begin();
+        game.font.draw(game.batch, "Welcome to Fruits!!! ", 100, 150);
+        game.font.draw(game.batch, "Tap anywhere to begin!", 100, 100);
+        game.batch.end();
+
+        if (Gdx.input.isTouched()) {
+            game.setScreen(new GameScreen(game));
+            dispose();
+        }
     }
     public void show() {
-
-    }
-    public void render(float delta) {
-
     }
     public void resize(int width, int height) {
     }
